fix a bug about unittest bound method
When using unittest and specify a method path from command line,
unittest will auto get that method.
If it's a <bound method>, unittest will call it as a test.

for example:
> python -m unittest tests.test_url.TestInclude.test_include_url

In former implementation, a <bound method> after __getattribute__ is no longer bounded.
That will cause unittest treat that method in a wrong way.

So I wrap coroutine function by metaclass instread of __getattribute__.